+++ sliceutil/example_test.go
@@ -13,6 +13,30 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+func ExampleCopy() {
+	s1 := []string{"A", "B", "C"}
+	s2 := Copy(s1)
+
+	fmt.Printf("%v", s2)
+	// Output: [A B C]
+}
+
+func ExampleCopyInts() {
+	s1 := []int{1, 2, 3}
+	s2 := CopyInts(s1)
+
+	fmt.Printf("%v", s2)
+	// Output: [1 2 3]
+}
+
+func ExampleCopyFloats() {
+	s1 := []float64{1.0, 5.0}
+	s2 := CopyFloats(s1)
+
+	fmt.Printf("%v", s2)
+	// Output: [1 5]
+}
+
 func ExampleStringToInterface() {
 	s := []string{"A", "B"}
 
@@ -46,7 +70,7 @@ func ExampleIndex() {
 func ExampleExclude() {
 	s := []string{"A", "B", "C", "D"}
 
-	fmt.Println(Exclude(s, []string{"B", "D"}))
+	fmt.Println(Exclude(s, "B", "D"))
 	// Output: [A C]
 }